Network Stuff
So it's a work in progress, haven't really had time to work on it too
much but I'm learning as much as I can as I go along.

-Players Spawn where I want them to.
-What I will do next is spawn both players at (0,0) without a sprite
then from there child all a the buttons at -1 or 1 away from the origin
and then the bars to the buttons with one beat for each that all have
there individual network ID's and Network Transforms to be able to sync.

-if you want to test it out
1 build it
2 play it in a window
3 play the scene in the editor, click lan host (first button)
4 go to the built version of the game and click the second button bellow
the first.